requirements.txt
jupyter
folium
requests
geojson
pyephem
pandas
matplotlib
argparse
